refactor(stress.base): convert stress killed to warning
since we have lots of cases when stopping the stress command
by the test is generating extra critical or error events

since we know the exact return code when this happens
we can convert those events to warning.

currently only implemented for stress command that uses
ContinuousEvent style, and the rest would need more testing
to be implmented.
