+++ sdcm/ndbench_thread.py
@@ -22,8 +22,8 @@ from sdcm.prometheus import nemesis_metrics_obj
 from sdcm.sct_events.loaders import NdBenchStressEvent, NDBENCH_ERROR_EVENTS_PATTERNS
 from sdcm.utils.common import FileFollowerThread
 from sdcm.utils.docker_remote import RemoteDocker
-from sdcm.stress_thread import format_stress_cmd_error, DockerBasedStressThread
-
+from sdcm.stress_thread import DockerBasedStressThread
+from sdcm.stress.base import format_stress_cmd_error
 
 LOGGER = logging.getLogger(__name__)
